:recycle: Refactor StatusBar
